fix: unblock trading + fix funding display + USTECH100 + score panel
- coherence thresholds lowered to 1.0 (SoDEX thin market max is ~1.37)
  - config: min_coherence + live_min_coherence 2.0 → 1.0
  - signal_generator Fallback 2: >= 2.0 → >= 1.0 (unblocks trade_direction)
  - coherence.py get_size_multiplier: starts at 1.0 (was 2.0)
- funding radar: remove synthetic rate substitution that inflated rates by
  4000x (1.25e-05 real → 0.05 synthetic → 5% display; now shows real rate)
- terminal asset panel: use interpreter.get_market_state() for Score column
  (market_engine states lacked weighted_score; interpreter has it)
- USTECH100 dead feed: route candle/OB/trade stores through SoDEX feed in
  hybrid mode (Bybit doesn't carry USTECH100; was getting empty buffers)
- schemas.py: coherence_score int → float (type mismatch vs MarketState)
- import BYBIT_SYMBOL_MAP in main.py for SoDEX-only asset routing

66/66 tests passing.

+++ execution/schemas.py
@@ -19,7 +19,7 @@ class TradeCandidate:
     initial_margin: float
     leverage: int
     rr_ratio: float
-    coherence_score: int
+    coherence_score: float
     size_multiplier: float
     signal_reason: str
     invalidation: str
